requirements.txt
psycopg2==2.8.4
python-dotenv==0.12.0
SQLAlchemy==1.3.15
